Clear all core warnings and turn on -Werror
The previous two commits enabled -Wall -Wextra on opensyde_core and fixed the
PCH mismatch, leaving 27 warnings documented as debt. Leaving them was the wrong
call: known noise is exactly what let the -Wparentheses on C_OscViewFiler's
inverted comparisons go unread for months. So they are cleared rather than
tracked.

26 -Wsign-compare, all the same shape -- an int32_t loop counter against .size().
Fixed by casting the size, not by retyping the counters: under the STW s32_/u32_
naming convention that would have meant renaming each variable and touching every
use in its body, which is far more churn for the same result.

Two of those were not loop bounds and were worth more than a warning fix:

    if (s32_Area == (orc_HexDataDump.at_Blocks.size() - 1))
    if (s32_Section < (mc_Sections.size() - 1))

On an empty container size() - 1 underflows to SIZE_MAX. Casting to int32_t
before subtracting makes it -1, so those comparisons are now correct as well as
quiet.

1 -Wunused-variable: an outer int32_t s32_Line shadowed by the loop's own
declaration nine lines later and never read.

Core now builds with zero diagnostics, so -Werror is on and this bug class cannot
reach a commit again. If it ever blocks legitimate work the fix is to silence the
specific warning at its site with a reason, not to drop the flag -- noted in the
CMake comment.

Verified: core clean with --clean-first, ctest 223/223, and all eight tools build
(BUILD_RC=0) -- necessary because these flags sit on a target whose PCH and
defines are PUBLIC.

+++ libraries/opensyde_core/can_dispatcher/dispatcher/C_CanDispatcher.cpp
@@ -331,7 +331,7 @@ int32_t C_CanDispatcher::DispatchIncoming(void)
       if (c_Return == Errc::success)
       {
          s32_NumMessages++;
-         for (s32_Loop = 0; s32_Loop < mc_InstalledClients.size(); s32_Loop++)
+         for (s32_Loop = 0; s32_Loop < static_cast<int32_t>(mc_InstalledClients.size()); s32_Loop++)
          {
             if (mc_InstalledClients[s32_Loop].c_RXFilter.DoesMessagePass(t_Msg) == true)
             {
@@ -596,7 +596,7 @@ std::error_code C_CanDispatcher::CAN_Read_Msg(T_STWCAN_Msg_RX & orc_Message)
    if (c_Return == Errc::success)
    {
       //dispatch to installed clients:
-      for (s32_Loop = 0; s32_Loop < mc_InstalledClients.size(); s32_Loop++)
+      for (s32_Loop = 0; s32_Loop < static_cast<int32_t>(mc_InstalledClients.size()); s32_Loop++)
       {
          if (mc_InstalledClients[s32_Loop].c_RXFilter.DoesMessagePass(orc_Message) == true)
          {
